fix(linux): address review feedback for Cursor provider
- Gate macOS-only browser session helpers behind os(macOS)
- Use an empty browser import order on non-macOS Linux builds
- Move Linux Cursor coverage into TestsLinux
