+++ TestsLinux/CursorLinuxTests.swift
@@ -1,0 +1,52 @@
+import CodexBarCore
+import Foundation
+import Testing
+@testable import CodexBarCLI
+
+@Suite
+struct CursorLinuxTests {
+    @Test
+    func resolveDefaultDBPathHonorsXDGConfigHome() {
+        let path = CursorAppAuthStore.resolveDefaultDBPath(
+            home: "/home/test",
+            environment: ["XDG_CONFIG_HOME": "/custom/config"])
+        #expect(path == "/custom/config/Cursor/User/globalStorage/state.vscdb")
+    }
+
+    @Test
+    func resolveDefaultDBPathFallsBackToDotConfig() {
+        let path = CursorAppAuthStore.resolveDefaultDBPath(
+            home: "/home/test",
+            environment: [:])
+        #expect(path == "/home/test/.config/Cursor/User/globalStorage/state.vscdb")
+    }
+
+    @Test
+    func cursorAutoSourceDoesNotRequireMacOSWebSupport() {
+        #expect(!CodexBarCLI.sourceModeRequiresWebSupport(
+            .auto,
+            provider: .cursor,
+            settings: ProviderSettingsSnapshot.make(
+                cursor: .init(cookieSource: .auto, manualCookieHeader: nil))))
+    }
+
+    @Test
+    func cursorManualCookieDoesNotRequireMacOSWebSupport() {
+        #expect(!CodexBarCLI.sourceModeRequiresWebSupport(
+            .web,
+            provider: .cursor,
+            settings: ProviderSettingsSnapshot.make(
+                cursor: .init(
+                    cookieSource: .manual,
+                    manualCookieHeader: "WorkosCursorSessionToken=test"))))
+    }
+
+    @Test
+    func cursorOffStillRequiresMacOSWebSupport() {
+        #expect(CodexBarCLI.sourceModeRequiresWebSupport(
+            .auto,
+            provider: .cursor,
+            settings: ProviderSettingsSnapshot.make(
+                cursor: .init(cookieSource: .off, manualCookieHeader: nil))))
+    }
+}
